requirements.txt
# AI Combat Platform - Dependencies
# Python 3.12

# Simulation
jsbsim==1.2.3

# Behavior Tree
py_trees>=2.4.0

# Machine Learning
torch>=2.0.0
numpy>=1.24.0
wandb>=0.16.0

# Environment
gymnasium>=0.29.0
gym>=0.26.0
pymap3d>=3.0.0
geographiclib>=2.0

# Web Backend
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
pyyaml>=6.0
pydantic>=2.0.0
icecream>=2.1.0
setproctitle>=1.3.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
cython>=3.0.0

# Visualization
matplotlib>=3.5.1
